% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootVar_function.R
\name{BootVar}
\alias{BootVar}
\title{Bootstrap inference in CRTs with recruitment bias}
\usage{
BootVar(
  B,
  Xobs,
  Zobs,
  Yobs,
  IDobs,
  treat_prop,
  fix_trt_num = TRUE,
  prop_compl = NULL,
  defiers = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{B}{Number of bootstrap samples.}

\item{Xobs}{Matrix with covariates for the recruited individuals. Rows
correspond to the individual, and columns to the different covariate. The
first column is a column of 1s for the intercept of the model.}

\item{Zobs}{Vector of treatments for the recruited individuals.}

\item{Yobs}{Vector of outcomes for the recruited individuals.}

\item{IDobs}{Vector of the cluster indices for the recruited individuals.
IDobs should include integers from 1 up to the number of different clusters.}

\item{treat_prop}{Proportion of clusters that are treated.}

\item{fix_trt_num}{Logical. Whether we want to keep the number of treated
clusters fixed in the bootstrap samples. This should agree with the
treatment assignment mechanism. Defaults to TRUE.}

\item{prop_compl}{The proportion of compliers. Bootstrap can be performed
without re-estimating the proportion of compliers within each bootstrap
sample. Then, set prop_compl to the estimated value. Defaults to NULL. If
left NULL, this will not be calculated. We suggest re-estimating the
proportion of compliers based on the bootstrap sample (leaving this to
NULL). In simulations, we have found similar inferential performance whether
this is re-estimated or not.}

\item{defiers}{Logical. Whether we allow for defiers to exist. If defiers
exist, we do not estimate effects on the compliers. Defaults to FALSE.}

\item{verbose}{Logical. Whether to print bootstrap progression. Defaults to
TRUE.}
}
\description{
Resamples clusters of the CRT and re-estimates quantities of interest in
order to acquire bootstrap-based inference.
}
