% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplierEffect_function.R
\name{ComplierEffect}
\alias{ComplierEffect}
\title{Estimation and inference on the complier subpopulation}
\usage{
ComplierEffect(prop_compl, effect_estimates, asym_var_est = NULL)
}
\arguments{
\item{prop_compl}{The proportion of compliers among the always-recruited and
the complier recruited. Numeric.}

\item{effect_estimates}{The effect estimates for the always-recruited and
the always+complier population. Vector of length 2.}

\item{asym_var_est}{2 x 2 matrix for the asymptotic variance of the causal 
estimators for the effect on the always and the always+complier populations.
Defaults to NULL. If null, inference will not be performed, and only the
estimate will be returned.}
}
\description{
Function that receives quantities on the always+complier population and the
always population, and returns an estimate of the causal effect and a
confidence interval for the complier population. It can only be used in the
absence of defiers.
}
