% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstWPSpars_function.R
\name{EstWPSpars}
\alias{EstWPSpars}
\title{Working propensity score estimation.}
\usage{
EstWPSpars(Xobs, Zobs, IDobs, treat_prop, inference = TRUE)
}
\arguments{
\item{Xobs}{Matrix with covariates for the recruited individuals. Rows
correspond to the individual, and columns to the different covariate. The
first column is a column of 1s for the intercept of the model.}

\item{Zobs}{Vector of treatment indicators for the recruited individuals.}

\item{IDobs}{Vector of integers with the index of the cluster to which the
recruited individuals belong. IDobs should include integers from 1 up to
the number of different clusters.}

\item{treat_prop}{Numeric. Proportion of clusters that are treated.}

\item{inference}{Logical. Whether the function should calculate the estimate
of the asymptotic variance for the working propensity score parameters.
Defaults to TRUE.}
}
\description{
Estimating the parameters of the working propensity score model which
expresses the conditional probability of treatment in the recruited sample.
}
