% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_function.R
\name{log_lik}
\alias{log_lik}
\title{Working propensity score log likelihood.}
\usage{
log_lik(theta, Xmat, Zvec, r)
}
\arguments{
\item{theta}{The parameters of the working propensity score model.}

\item{Xmat}{The covariate matrix where rows correspond to individuals and
columns to covariates. The first column is equal to 1 for the intercept.}

\item{Zvec}{The observed treatment of the units.}

\item{r}{Numeric. Corresponds to the ratio of the treatment probability over
1 - treatment probability, where treatment probability is the probability
used to assign the cluster-level treatment.}
}
\description{
Function that calculates the log-likelihood of the treatment among the
recruited according to the working propensity score model.
}
