% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimultEstimateCRT_function.R
\name{SimultEstimateCRT}
\alias{SimultEstimateCRT}
\title{Estimation and inference of causal effects in cluster randomized trials}
\usage{
SimultEstimateCRT(
  Zobs,
  Yobs,
  weights,
  IDobs,
  Xobs = NULL,
  ps = c("known", "estimated"),
  est_wps_pars = NULL,
  inference = TRUE,
  treat_prop = 1/2
)
}
\arguments{
\item{Zobs}{Vector of treatments for the recruited individuals.}

\item{Yobs}{Vector of outcomes for the recruited individuals.}

\item{weights}{Matrix of weights for each individual used for estimation.
Columns correspond to weights for the always-recruited, for the always+
compliers, and for the recruited.}

\item{IDobs}{Vector of the cluster indices for the recruited individuals.}

\item{ps}{Whether the propensity score is known or estimated. Options are
'known' or 'estimated'. Defaults to 'known'.}

\item{est_wps_pars}{If the PS is estimated, est_wps_pars is needed. This
is the output of the EstWPSpars function used to estimate the parameters.
This output will include quantities needed for variance estimation in the
A and V matrices.}

\item{inference}{Logical. Whether asymptotic variance and inference is
performed. Defaults to TRUE.}

\item{treat_prop}{Proportion of clusters that are treated. Defaults to 0.5.}
}
\description{
Estimating the causal effect on the subpopulation of the always-recruited
and defiers, the always- and the complier-recruited, and the recruited.
}
\details{
Builds on the NewEstimateCRT function that does estimation and inference for
always+defiers and recruited only.
}
